backend/requirements.txt
Flask
flask_cors
python-dotenv
gunicorn
